start to refactor create trip form for create trip function, not yet finished
