+++ screens/ActivePongs.js
@@ -1,41 +1,41 @@
-import React from "react"
-import { View, StyleSheet, Text, TouchableHighlight } from "react-native"
-import { useSelector, useDispatch } from "react-redux"
-import { cancelRequest } from "modules/tripActions"
-import { Icon } from "react-native-elements"
-import Receipt from "screens/Receipt"
+import React from "react";
+import { View, StyleSheet, Text, TouchableHighlight } from "react-native";
+import { useSelector, useDispatch } from "react-redux";
+import { cancelRequest } from "modules/tripActions";
+import { Icon } from "react-native-elements";
+import Receipt from "screens/Receipt";
 
 const ActivePongs = () => {
-  const dispatch = useDispatch()
-  const myPong = useSelector((state) => state.myPong)
+  const dispatch = useDispatch();
+  const myPong = useSelector((state) => state.myPong);
   const cancelledRequestResponse = useSelector(
     (state) => state.cancelledRequestResponse
-  )
+  );
 
-  const pongStatus = myPong.status
+  const pongStatus = myPong.status;
 
-  let statusColor
+  let statusColor;
   if (pongStatus === "pending") {
     statusColor = (
       <>
         <Text style={styles.status}>Your request is </Text>
         <Text style={styles.pending}>{pongStatus}</Text>
       </>
-    )
+    );
   } else if (pongStatus === "accepted") {
     statusColor = (
       <>
         <Text style={styles.status}>Your request is </Text>
         <Text style={styles.accepted}>{pongStatus}</Text>
       </>
-    )
+    );
   } else {
     statusColor = (
       <>
         <Text style={styles.status}>Your request is </Text>
         <Text style={styles.rejected}>{pongStatus}</Text>
       </>
-    )
+    );
   }
 
   return (
@@ -47,11 +47,11 @@ const ActivePongs = () => {
         <Text style={styles.item}>{myPong.item1}</Text>
       </View>
       <View style={styles.itemContainer}>
-        {myPong.item2 !== '' && <Icon name="ios-cart" type="ionicon" />}
+        {myPong.item2 !== "" && <Icon name="ios-cart" type="ionicon" />}
         <Text style={styles.item}>{myPong.item2}</Text>
       </View>
       <View style={styles.itemContainer}>
-        {myPong.item3 !== '' && <Icon name="ios-cart" type="ionicon" />}
+        {myPong.item3 !== "" && <Icon name="ios-cart" type="ionicon" />}
         <Text style={styles.item}>{myPong.item3}</Text>
       </View>
       <Receipt />
@@ -62,7 +62,7 @@ const ActivePongs = () => {
           <TouchableHighlight
             style={styles.cancelButton}
             onPress={() => {
-              cancelRequest(myPong.id, dispatch)
+              cancelRequest(myPong.id, dispatch);
             }}
           >
             <Text id={"cancel-button"} style={styles.requestButtonText}>
@@ -72,8 +72,8 @@ const ActivePongs = () => {
         )}
       </View>
     </View>
-  )
-}
+  );
+};
 
 const styles = StyleSheet.create({
   pong: {
@@ -83,7 +83,7 @@ const styles = StyleSheet.create({
     backgroundColor: "white",
     shadowColor: "black",
     shadowOpacity: 2.0,
-    height: 350
+    height: 350,
   },
   item: {
     fontSize: 18,
@@ -153,6 +153,6 @@ const styles = StyleSheet.create({
     fontWeight: "normal",
     fontFamily: "Futura-Medium",
   },
-})
+});
 
-export default ActivePongs
+export default ActivePongs;
